refactor exception handler from messengerplatform package
