+++ src/main/java/com/xalpol12/messengerbot/messengerplatform/exception/customexception/IncorrectWebhookModeException.java
@@ -1,4 +1,4 @@
-package com.xalpol12.messengerbot.messengerplatform.exception;
+package com.xalpol12.messengerbot.messengerplatform.exception.customexception;
 
 /**
  * Exception occurs when received webhook mode is not the
