+++ src/test/java/com/xalpol12/messengerbot/messengerplatform/exception/MessengerPlatformExceptionHandlerTest.java
@@ -1,5 +1,10 @@
 package com.xalpol12.messengerbot.messengerplatform.exception;
 
+import com.xalpol12.messengerbot.core.exception.response.CustomErrorResponse;
+import com.xalpol12.messengerbot.messengerplatform.exception.customexception.IncorrectTokenException;
+import com.xalpol12.messengerbot.messengerplatform.exception.customexception.IncorrectWebhookModeException;
+import com.xalpol12.messengerbot.messengerplatform.exception.customexception.IncorrectWebhookObjectTypeException;
+import com.xalpol12.messengerbot.messengerplatform.exception.customexception.RequestSignatureValidationException;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
@@ -22,7 +27,7 @@ class MessengerPlatformExceptionHandlerTest {
         String message = "Exception message";
         IncorrectWebhookModeException exception = new IncorrectWebhookModeException(message);
 
-        ResponseEntity<String> response = exceptionHandler.handleWebhookVerificationExceptions(exception);
+        ResponseEntity<CustomErrorResponse> response = exceptionHandler.handleWebhookVerificationExceptions(exception);
 
         assertAll(() -> {
             assertEquals(HttpStatus.FORBIDDEN, response.getStatusCode());
@@ -35,7 +40,7 @@ class MessengerPlatformExceptionHandlerTest {
         String message = "Exception message";
         IncorrectTokenException exception = new IncorrectTokenException(message);
 
-        ResponseEntity<String> response = exceptionHandler.handleWebhookVerificationExceptions(exception);
+        ResponseEntity<CustomErrorResponse> response = exceptionHandler.handleWebhookVerificationExceptions(exception);
 
         assertAll(() -> {
             assertEquals(HttpStatus.FORBIDDEN, response.getStatusCode());
@@ -48,7 +53,7 @@ class MessengerPlatformExceptionHandlerTest {
         String message = "Exception message";
         RequestSignatureValidationException exception = new RequestSignatureValidationException(message);
 
-        ResponseEntity<String> response = exceptionHandler.handleHashValidationException(exception);
+        ResponseEntity<CustomErrorResponse> response = exceptionHandler.handleHashValidationException(exception);
 
         assertAll(() -> {
             assertEquals(HttpStatus.BAD_REQUEST, response.getStatusCode());
@@ -61,7 +66,7 @@ class MessengerPlatformExceptionHandlerTest {
         String message = "Exception message";
         IncorrectWebhookObjectTypeException exception = new IncorrectWebhookObjectTypeException(message);
 
-        ResponseEntity<String> response = exceptionHandler.handleIncorrectWebhookObjectType(exception);
+        ResponseEntity<CustomErrorResponse> response = exceptionHandler.handleIncorrectWebhookObjectType(exception);
 
         assertAll(() -> {
             assertEquals(HttpStatus.NOT_FOUND, response.getStatusCode());
