+++ src/test/java/com/xalpol12/messengerbot/messengerplatform/service/WebhookServiceTest.java
@@ -2,8 +2,8 @@ package com.xalpol12.messengerbot.messengerplatform.service;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.xalpol12.messengerbot.messengerplatform.config.secrets.SecretsConfig;
-import com.xalpol12.messengerbot.messengerplatform.exception.IncorrectTokenException;
-import com.xalpol12.messengerbot.messengerplatform.exception.IncorrectWebhookModeException;
+import com.xalpol12.messengerbot.messengerplatform.exception.customexception.IncorrectTokenException;
+import com.xalpol12.messengerbot.messengerplatform.exception.customexception.IncorrectWebhookModeException;
 import com.xalpol12.messengerbot.publisher.service.SubscriberService;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
